refactor: move vancouver date processing into own module
